+++ malachite-q/src/test_util/generators/random.rs
@@ -113,6 +113,17 @@ pub fn random_gaussian_rational_gen_var_2(config: &GenConfig) -> It<GaussianRati
     ))
 }
 
+pub fn random_gaussian_rational_gen_var_3(config: &GenConfig) -> It<GaussianRational> {
+    Box::new(
+        random_gaussian_rationals(
+            EXAMPLE_SEED,
+            config.get_or("mean_bits_n", 64),
+            config.get_or("mean_bits_d", 1),
+        )
+        .filter(|x| x.real != 0u32 || x.imaginary != 0u32),
+    )
+}
+
 // -- (GaussianRational, GaussianInteger) --
 
 pub fn random_gaussian_rational_gaussian_integer_pair_gen(