+++ malachite-q/src/test_util/generators/special_random.rs
@@ -129,6 +129,19 @@ pub fn special_random_gaussian_rational_gen_var_2(config: &GenConfig) -> It<Gaus
     ))
 }
 
+pub fn special_random_gaussian_rational_gen_var_3(config: &GenConfig) -> It<GaussianRational> {
+    Box::new(
+        striped_random_gaussian_rationals(
+            EXAMPLE_SEED,
+            config.get_or("mean_stripe_n", 32),
+            config.get_or("mean_stripe_d", 1),
+            config.get_or("mean_bits_n", 64),
+            config.get_or("mean_bits_d", 1),
+        )
+        .filter(|x| x.real != 0u32 || x.imaginary != 0u32),
+    )
+}
+
 // -- (GaussianRational, GaussianInteger) --
 
 pub fn special_random_gaussian_rational_gaussian_integer_pair_gen(